requirements.txt
requests
facebook-sdk
prettytable
networkx